typescrpit error fix test
