+++ apps/docs/app/global.css
@@ -10,13 +10,17 @@ body {
 
 @theme {
   --color-fd-primary: #33c482; /* Green from Sim logo */
-  --font-geist-sans: var(--font-geist-sans);
-  --font-geist-mono: var(--font-geist-mono);
+  --font-geist-sans: "Inter", ui-sans-serif, system-ui, sans-serif;
+  --font-geist-mono: "Geist Mono", ui-monospace, monospace;
 }
 
-/* Ensure primary color is set in both light and dark modes */
+/* Ensure primary color and font variables (loaded via link in layout, not next/font) */
 :root {
   --color-fd-primary: #33c482;
+  --font-geist-sans: "Inter", ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
+    "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+  --font-geist-mono: "Geist Mono", ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas,
+    "Liberation Mono", "Courier New", monospace;
 }
 
 .dark {
