+++ apps/sim/app/api/chat/route.ts
@@ -309,7 +309,7 @@ export async function POST(request: NextRequest) {
       try {
         const { PlatformEvents } = await import('@/lib/core/telemetry')
         PlatformEvents.chatDeployed({
-          chatId: id,
+          chatId,
           workflowId,
           authType,
           hasOutputConfigs: outputConfigs.length > 0,
