+++ apps/sim/app/api/files/utils.ts
@@ -160,20 +160,18 @@ function sanitizeFilename(filename: string): string {
 
 export function findLocalFile(filename: string): string | null {
   try {
-    // Handle agent-generated-images path directly (no sanitization needed)
     if (filename.startsWith('agent-generated-images/')) {
-      // Store at /apps/sim/agent-generated-images
-      const appPath = resolve(process.cwd(), filename)
+      if (filename.includes('..')) {
+        return null
+      }
       const appAllowedDir = resolve(process.cwd(), 'agent-generated-images')
-      
-      // Must be within allowed directory but NOT the directory itself
+      const appPath = resolve(process.cwd(), filename)
       const isWithinAppDir =
         appPath.startsWith(appAllowedDir + sep) && appPath !== appAllowedDir
 
       if (isWithinAppDir && existsSync(appPath)) {
         return appPath
       }
-      
       return null
     }
