+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/chat/components/start-block-input-modal/start-block-input-modal.tsx
@@ -290,7 +290,7 @@ export function StartBlockInputModal({
                 {fieldType === 'boolean' ? (
                   <div className='flex items-center gap-2'>
                     <input
-                      ref={isFirstField ? (el) => (firstInputRef.current = el) : undefined}
+                      ref={isFirstField ? (el) => { firstInputRef.current = el } : undefined}
                       type='checkbox'
                       id={fieldName}
                       checked={value === true || value === 'true'}
@@ -303,7 +303,7 @@ export function StartBlockInputModal({
                   </div>
                 ) : fieldType === 'number' ? (
                   <Input
-                    ref={isFirstField ? (el) => (firstInputRef.current = el) : undefined}
+                    ref={isFirstField ? (el) => { firstInputRef.current = el } : undefined}
                     id={fieldName}
                     type='number'
                     value={
@@ -352,7 +352,7 @@ export function StartBlockInputModal({
                   />
                 ) : fieldType === 'object' || fieldType === 'array' ? (
                   <textarea
-                    ref={isFirstField ? (el) => (firstInputRef.current = el) : undefined}
+                    ref={isFirstField ? (el) => { firstInputRef.current = el } : undefined}
                     id={fieldName}
                     value={
                       typeof value === 'string'
@@ -380,7 +380,7 @@ export function StartBlockInputModal({
                   />
                 ) : (
                   <Input
-                    ref={isFirstField ? (el) => (firstInputRef.current = el) : undefined}
+                    ref={isFirstField ? (el) => { firstInputRef.current = el } : undefined}
                     id={fieldName}
                     type='text'
                     value={(() => {
