+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/panel/components/editor/components/sub-block/components/slack-channel-selector/index.tsx
@@ -59,10 +59,10 @@ export function SlackChannelSelector({
 
   const { finalDisabled } = useDependsOnGate(
     blockId,
-    { dependsOn },
+    { id: subBlockId, type: 'channel-selector', dependsOn },
     {
       isPreview,
-      subBlockValues,
+      previewContextValues: subBlockValues,
       disabled,
     }
   )
