+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/panel/components/editor/components/sub-block/hooks/use-depends-on-gate.ts
@@ -177,5 +177,6 @@ export function useDependsOnGate(
     depsSatisfied,
     blocked,
     finalDisabled,
+    dependencyValues: dependencyValuesMap,
   }
 }
