+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/panel/components/editor/components/sub-block/sub-block.tsx
@@ -554,7 +554,9 @@ function SubBlockComponent({
                 subBlockId={config.id}
                 placeholder={config.placeholder}
                 isPreview={isPreview}
-                previewValue={previewValue}
+                previewValue={
+                  Array.isArray(previewValue) ? (previewValue[0] as string | undefined) : (previewValue ?? undefined)
+                }
                 disabled={isDisabled}
               />
             </div>
@@ -566,11 +568,14 @@ function SubBlockComponent({
               blockId={blockId}
               subBlockId={config.id}
               options={config.options as { label: string; id: string }[]}
-              defaultValue={
-                typeof config.value === 'function'
-                  ? config.value(subBlockValues || {})
-                  : config.value
-              }
+              defaultValue={(() => {
+                const raw =
+                  typeof config.value === 'function'
+                    ? config.value(subBlockValues || {})
+                    : config.value
+                const v = Array.isArray(raw) ? (raw[0] as string | undefined) ?? undefined : raw
+                return v === null ? undefined : v
+              })()}
               placeholder={config.placeholder}
               isPreview={isPreview}
               previewValue={previewValue}
@@ -1068,7 +1073,11 @@ function SubBlockComponent({
             isPreview={isPreview}
             subBlockValues={subBlockValues}
             disabled={isDisabled}
-            dependsOn={config.dependsOn}
+            dependsOn={
+              Array.isArray(config.dependsOn)
+                ? config.dependsOn
+                : [...(config.dependsOn?.all ?? []), ...(config.dependsOn?.any ?? [])]
+            }
           />
         )
       case 'arena-states-selector':
