+++ apps/sim/lib/core/security/input-validation.ts
@@ -1,5 +1,3 @@
-import type { AgentOptions, RequestOptions } from 'http'
-import type { LookupFunction } from 'net'
 import { createLogger } from '@sim/logger'
 import * as ipaddr from 'ipaddr.js'
 
@@ -748,7 +746,7 @@ export function validateProxyUrl(
  * - IPv4-mapped IPv6 (::ffff:127.0.0.1)
  * - Various edge cases that regex patterns miss
  */
-function isPrivateOrReservedIP(ip: string): boolean {
+export function isPrivateOrReservedIP(ip: string): boolean {
   try {
     if (!ipaddr.isValid(ip)) {
       return true
@@ -771,65 +769,6 @@ export interface AsyncValidationResult extends ValidationResult {
   originalHostname?: string
 }
 
-/**
- * Validates a URL and resolves its DNS to prevent SSRF via DNS rebinding
- *
- * This function:
- * 1. Performs basic URL validation (protocol, format)
- * 2. Resolves the hostname to an IP address
- * 3. Validates the resolved IP is not private/reserved
- * 4. Returns the resolved IP for use in the actual request
- *
- * @param url - The URL to validate
- * @param paramName - Name of the parameter for error messages
- * @returns AsyncValidationResult with resolved IP for DNS pinning
- */
-export async function validateUrlWithDNS(
-  url: string | null | undefined,
-  paramName = 'url'
-): Promise<AsyncValidationResult> {
-  const basicValidation = validateExternalUrl(url, paramName)
-  if (!basicValidation.isValid) {
-    return basicValidation
-  }
-
-  const parsedUrl = new URL(url!)
-  const hostname = parsedUrl.hostname
-
-  try {
-    // Dynamic import to avoid client-side bundle issues
-    const dns = await import('dns/promises')
-    const { address } = await dns.lookup(hostname)
-
-    if (isPrivateOrReservedIP(address)) {
-      logger.warn('URL resolves to blocked IP address', {
-        paramName,
-        hostname,
-        resolvedIP: address,
-      })
-      return {
-        isValid: false,
-        error: `${paramName} resolves to a blocked IP address`,
-      }
-    }
-
-    return {
-      isValid: true,
-      resolvedIP: address,
-      originalHostname: hostname,
-    }
-  } catch (error) {
-    logger.warn('DNS lookup failed for URL', {
-      paramName,
-      hostname,
-      error: error instanceof Error ? error.message : String(error),
-    })
-    return {
-      isValid: false,
-      error: `${paramName} hostname could not be resolved`,
-    }
-  }
-}
 export interface SecureFetchOptions {
   method?: string
   headers?: Record<string, string>
@@ -870,162 +809,6 @@ export interface SecureFetchResponse {
   text: () => Promise<string>
   json: () => Promise<unknown>
   arrayBuffer: () => Promise<ArrayBuffer>
-}
-
-const DEFAULT_MAX_REDIRECTS = 5
-
-function isRedirectStatus(status: number): boolean {
-  return status >= 300 && status < 400 && status !== 304
-}
-
-function resolveRedirectUrl(baseUrl: string, location: string): string {
-  try {
-    return new URL(location, baseUrl).toString()
-  } catch {
-    throw new Error(`Invalid redirect location: ${location}`)
-  }
-}
-
-/**
- * Performs a fetch with IP pinning to prevent DNS rebinding attacks.
- * Uses the pre-resolved IP address while preserving the original hostname for TLS SNI.
- * Follows redirects securely by validating each redirect target.
- */
-export async function secureFetchWithPinnedIP(
-  url: string,
-  resolvedIP: string,
-  options: SecureFetchOptions = {},
-  redirectCount = 0
-): Promise<SecureFetchResponse> {
-  const maxRedirects = options.maxRedirects ?? DEFAULT_MAX_REDIRECTS
-
-  // Dynamic imports to avoid client-side bundle issues
-  const [httpModule, httpsModule] = await Promise.all([import('http'), import('https')])
-  // Type assertions for Node.js built-in modules
-  const http = httpModule as unknown as typeof import('http')
-  const https = httpsModule as unknown as typeof import('https')
-
-  return new Promise((resolve, reject) => {
-    const parsed = new URL(url)
-    const isHttps = parsed.protocol === 'https:'
-    const defaultPort = isHttps ? 443 : 80
-    const port = parsed.port ? Number.parseInt(parsed.port, 10) : defaultPort
-
-    const isIPv6 = resolvedIP.includes(':')
-    const family = isIPv6 ? 6 : 4
-
-    const lookup: LookupFunction = (_hostname, options, callback) => {
-      if (options.all) {
-        callback(null, [{ address: resolvedIP, family }])
-      } else {
-        callback(null, resolvedIP, family)
-      }
-    }
-
-    const agentOptions: AgentOptions = { lookup }
-
-    const agent = isHttps ? new https.Agent(agentOptions) : new http.Agent(agentOptions)
-
-    // Remove accept-encoding since Node.js http/https doesn't auto-decompress
-    // Headers are lowercase due to Web Headers API normalization in executeToolRequest
-    const { 'accept-encoding': _, ...sanitizedHeaders } = options.headers ?? {}
-
-    const requestOptions: RequestOptions = {
-      hostname: parsed.hostname,
-      port,
-      path: parsed.pathname + parsed.search,
-      method: options.method || 'GET',
-      headers: sanitizedHeaders,
-      agent,
-      timeout: options.timeout || 30000,
-    }
-
-    const protocol = isHttps ? https : http
-    const req = protocol.request(requestOptions, (res) => {
-      const statusCode = res.statusCode || 0
-      const location = res.headers.location
-
-      if (isRedirectStatus(statusCode) && location && redirectCount < maxRedirects) {
-        res.resume()
-        const redirectUrl = resolveRedirectUrl(url, location)
-
-        validateUrlWithDNS(redirectUrl, 'redirectUrl')
-          .then((validation) => {
-            if (!validation.isValid) {
-              reject(new Error(`Redirect blocked: ${validation.error}`))
-              return
-            }
-            return secureFetchWithPinnedIP(
-              redirectUrl,
-              validation.resolvedIP!,
-              options,
-              redirectCount + 1
-            )
-          })
-          .then((response) => {
-            if (response) resolve(response)
-          })
-          .catch(reject)
-        return
-      }
-
-      if (isRedirectStatus(statusCode) && location && redirectCount >= maxRedirects) {
-        res.resume()
-        reject(new Error(`Too many redirects (max: ${maxRedirects})`))
-        return
-      }
-
-      const chunks: Buffer[] = []
-
-      res.on('data', (chunk: Buffer) => chunks.push(chunk))
-
-      res.on('error', (error) => {
-        reject(error)
-      })
-
-      res.on('end', () => {
-        const bodyBuffer = Buffer.concat(chunks)
-        const body = bodyBuffer.toString('utf-8')
-        const headersRecord: Record<string, string> = {}
-        for (const [key, value] of Object.entries(res.headers)) {
-          if (typeof value === 'string') {
-            headersRecord[key.toLowerCase()] = value
-          } else if (Array.isArray(value)) {
-            headersRecord[key.toLowerCase()] = value.join(', ')
-          }
-        }
-
-        resolve({
-          ok: statusCode >= 200 && statusCode < 300,
-          status: statusCode,
-          statusText: res.statusMessage || '',
-          headers: new SecureFetchHeaders(headersRecord),
-          text: async () => body,
-          json: async () => JSON.parse(body),
-          arrayBuffer: async () =>
-            bodyBuffer.buffer.slice(
-              bodyBuffer.byteOffset,
-              bodyBuffer.byteOffset + bodyBuffer.byteLength
-            ),
-        })
-      })
-    })
-
-    req.on('error', (error) => {
-      reject(error)
-    })
-
-    req.on('timeout', () => {
-      req.destroy()
-      reject(new Error('Request timeout'))
-    })
-
-    if (options.body) {
-      req.write(options.body)
-    }
-
-    req.end()
-  })
 }
 
 /**
