+++ apps/sim/tools/index.ts
@@ -1,6 +1,5 @@
 import { createLogger } from '@sim/logger'
 import { generateInternalToken } from '@/lib/auth/internal'
-import { secureFetchWithPinnedIP, validateUrlWithDNS } from '@/lib/core/security/input-validation'
 import { generateRequestId } from '@/lib/core/utils/request'
 import { getBaseUrl } from '@/lib/core/utils/urls'
 import { parseMcpToolId } from '@/lib/mcp/utils'
@@ -673,6 +672,9 @@ async function executeToolRequest(
         body: requestParams.body,
       })
     } else {
+      const { validateUrlWithDNS, secureFetchWithPinnedIP } = await import(
+        '@/lib/core/security/input-validation.server'
+      )
       const urlValidation = await validateUrlWithDNS(fullUrl, 'toolUrl')
       if (!urlValidation.isValid) {
         throw new Error(`Invalid tool URL: ${urlValidation.error}`)
